#70 Our application should modularize blueprints components
* update readme to include information about how to perform a migration
* update blueprints to use modules that contain (api, dao, views, templates, etc)
* modularize auth and admin blueprints
* delete navigation.html code ( will restore it later)
