+++ simwire_plugin/blueprints/__init__.py
@@ -1,8 +1,3 @@
-from flask import Blueprint
-
-home_bp = Blueprint('home', __name__)
-admin_bp = Blueprint('admin', __name__)
-auth_bp = Blueprint('auth', __name__)
-from .home_blueprint import *
-from .auth_blueprint import *
-from .admin_blueprint import *
+from simwire_plugin.blueprints.admin.admin_views import *
+from simwire_plugin.blueprints.auth.auth_views import *
+from simwire_plugin.blueprints.home_blueprint import *
