+++ simwire_plugin/blueprints/admin/admin_api.py
@@ -1,0 +1,2 @@
+# Code here is for restful calls against the module.
+# The response from these api calls will ALWAYS be JSON.
