+++ simwire_plugin/blueprints/admin/admin_dao.py
@@ -1,0 +1,3 @@
+# code here is for access the database directly
+# The return type for these functions will always be representation of the model or the model itself.
+# Note: It is my expectation that this would only return a representation of the model.
