+++ simwire_plugin/blueprints/admin/templates/administration.html
