+++ simwire_plugin/blueprints/home_blueprint.py
@@ -1,10 +1,7 @@
-from flask import render_template, url_for
-from flask_login import LoginManager, login_required, current_user, logout_user, login_user
+from flask import render_template
+from flask import Blueprint
 
-from pathlib import Path
-from . import home_bp
-
-
+home_bp = Blueprint('home', __name__)
 @home_bp.route('/')
 def index():
-    return "this is the index page"
+    return render_template('homepage.html')
