+++ simwire_plugin/templates/navigation.html
@@ -1,39 +1,3 @@
-<!-- Fixed navbar
-    <nav class="navbar navbar-fixed-top">
-      <div class="container">
-        <div class="navbar-header">
-          <button type="button" class="navbar-toggle collapsed" data-toggle="collapse" data-target="#navbar" aria-expanded="false" aria-controls="navbar">
-            <span class="sr-only">Toggle navigation</span>
-            <span class="icon-bar"></span>
-            <span class="icon-bar"></span>
-            <span class="icon-bar"></span>
-          </button>
-          {#<a class="navbar-brand" href="">
-
-                {% block logo %}
-                <img src="" >
-                {% endblock %}
-          </a>#}
-        </div>
-        <div id="navbar" class="navbar-collapse collapse">
-          {% block user_menu %}
-          <ul class="nav navbar-nav navbar-right">
-            <li class=""><a href="{{url_for('admin_landing_page')}}">Admin</a></li>
-            <li class="dropdown">
-              <a href="#" class="dropdown-toggle" data-toggle="dropdown" role="button" aria-haspopup="true" aria-expanded="false">
-              My Account
-              <span class="caret"></span></a>
-              <ul class="dropdown-menu">
-                <li>My Account</li>
-                <li><a href="/logout">Logout</a></li>
-              </ul>
-            </li>
-          </ul>
-          {% endblock %}
-        </div>
-      </div>
-    </nav>
--->
 <nav class="navbar navbar-expand-lg navbar-dark bg-primary">
  <button class="navbar-toggler"
           type="button" data-toggle="collapse" data-target="#navbarNav"
@@ -43,10 +7,10 @@
   <div class="collapse navbar-collapse" id="navbarNav">
     <ul class="navbar-nav">
       <li class="nav-item active">
-        <a class="nav-link" href="{{url_for('homepage')}}">Home</a>
+        <a class="nav-link" href="{{url_for('home.index')}}">Home</a>
       </li>
       <li class="nav-item">
-        <a class="nav-link" href="{{url_for('admin_landing_page')}}">Admin</a>
+        <a class="nav-link" href="{{url_for('admin.admin_landing_page')}}">Admin</a>
       </li>
     </ul>
   </div>
